the new up file runs the updates for silverblue
Give the script executable permissions and add it on startup. The script will show a progress bar with the upgrade progress of the os and the apps, when it's complete will close. It need no sudo privileges (like the automatic updates setup guides) the UI is meant for KDE desktop.